requirements.txt
langchain
langgraph
langchain-openai
langchain-ollama
langchain-community
openai
fastapi[standard]
uvicorn
kafka-python
redis
langmem
langfuse
sqlalchemy
mcp
python-dotenv
langchain-core
dependency-injector
langchain-mcp-adapters
psycopg-pool 
psycopg2
psycopg 
langgraph-checkpoint-redis
aiokafka
